View fix

+++ Controllers/BaiVietsController.cs
@@ -1,9 +1,11 @@
-﻿using System.Linq;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using CuoiKiLTC.Models;
-using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace CuoiKiLTC.Controllers
 {
@@ -17,22 +19,59 @@ namespace CuoiKiLTC.Controllers
         }
 
         // GET: BaiViets
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string searchString)
         {
-            return View(await _context.BaiViets.Include(b => b.TheLoai).ToListAsync());
+            var baiViets = from bv in _context.BaiViets.Include(b => b.Admin).Include(b => b.TheLoai)
+                           select bv;
+
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                baiViets = baiViets.Where(bv => bv.TieuDe.Contains(searchString) ||
+                                                bv.NoiDung.Contains(searchString) ||
+                                                bv.TacGia.Contains(searchString) ||
+                                                bv.Admin.UserName.Contains(searchString) ||
+                                                bv.TheLoai.TenTheLoai.Contains(searchString));
+            }
+
+            ViewData["CurrentFilter"] = searchString;
+
+            return View(await baiViets.ToListAsync());
+        }
+
+        // GET: BaiViets/Details/5
+        public async Task<IActionResult> Details(int? id)
+        {
+            if (id == null || _context.BaiViets == null)
+            {
+                return NotFound();
+            }
+
+            var baiViet = await _context.BaiViets
+                .Include(b => b.Admin)
+                .Include(b => b.TheLoai)
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (baiViet == null)
+            {
+                return NotFound();
+            }
+
+            return View(baiViet);
         }
 
         // GET: BaiViets/Create
         public IActionResult Create()
         {
-            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "TenTheLoai");
+            ViewData["AdminId"] = new SelectList(_context.Admins, "Id", "Id");
+            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "Id");
             return View();
         }
 
         // POST: BaiViets/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,TieuDe,NoiDung,NgayDang,TacGia,TheLoaiId")] BaiViet baiViet)
+        public async Task<IActionResult> Create([Bind("Id,TieuDe,NoiDung,NgayDang,TacGia,TheLoaiId,AdminId")] BaiViet baiViet)
         {
             if (ModelState.IsValid)
             {
@@ -40,14 +79,15 @@ namespace CuoiKiLTC.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "TenTheLoai", baiViet.TheLoaiId);
+            ViewData["AdminId"] = new SelectList(_context.Admins, "Id", "Id", baiViet.AdminId);
+            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "Id", baiViet.TheLoaiId);
             return View(baiViet);
         }
 
         // GET: BaiViets/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.BaiViets == null)
             {
                 return NotFound();
             }
@@ -57,14 +97,17 @@ namespace CuoiKiLTC.Controllers
             {
                 return NotFound();
             }
-            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "TenTheLoai", baiViet.TheLoaiId);
+            ViewData["AdminId"] = new SelectList(_context.Admins, "Id", "Id", baiViet.AdminId);
+            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "Id", baiViet.TheLoaiId);
             return View(baiViet);
         }
 
         // POST: BaiViets/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,TieuDe,NoiDung,NgayDang,TacGia,TheLoaiId")] BaiViet baiViet)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,TieuDe,NoiDung,NgayDang,TacGia,TheLoaiId,AdminId")] BaiViet baiViet)
         {
             if (id != baiViet.Id)
             {
@@ -91,19 +134,21 @@ namespace CuoiKiLTC.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "TenTheLoai", baiViet.TheLoaiId);
+            ViewData["AdminId"] = new SelectList(_context.Admins, "Id", "Id", baiViet.AdminId);
+            ViewData["TheLoaiId"] = new SelectList(_context.TheLoais, "Id", "Id", baiViet.TheLoaiId);
             return View(baiViet);
         }
 
         // GET: BaiViets/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.BaiViets == null)
             {
                 return NotFound();
             }
 
             var baiViet = await _context.BaiViets
+                .Include(b => b.Admin)
                 .Include(b => b.TheLoai)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (baiViet == null)
@@ -119,15 +164,23 @@ namespace CuoiKiLTC.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.BaiViets == null)
+            {
+                return Problem("Entity set 'QuanLyCongTyContext.BaiViets'  is null.");
+            }
             var baiViet = await _context.BaiViets.FindAsync(id);
-            _context.BaiViets.Remove(baiViet);
+            if (baiViet != null)
+            {
+                _context.BaiViets.Remove(baiViet);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool BaiVietExists(int id)
         {
-            return _context.BaiViets.Any(e => e.Id == id);
+          return (_context.BaiViets?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
